backend/requirements.txt
# Для загрузки .env при локальном запуске (в Docker переменные передаются иначе)
python-dotenv>=1.0.0
